backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.6.1
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.13.3
openai==1.51.2
httpx==0.27.2
pandas==2.2.3
python-docx==1.1.2
openpyxl==3.1.5
PyPDF2==3.0.1
Pillow==10.4.0
python-dotenv==1.0.1
